+++ src/pyramid_bimt/hooks.py
@@ -5,7 +5,6 @@ from pyramid_bimt.models import User
 
 def get_authenticated_user(request):
     """Get the authenticated user for this ``request``, if there is one."""
-    if request:
-        email = request.unauthenticated_userid
-        if email:
-            return User.by_email(email)
+    email = request.unauthenticated_userid
+    if email:
+        return User.by_email(email)