+++ src/pyramid_bimt/views/auditlog.py
@@ -5,6 +5,7 @@ from colanderalchemy import SQLAlchemySchemaNode
 from collections import OrderedDict
 from datetime import datetime
 from pyramid.httpexceptions import HTTPFound
+from pyramid.security import ALL_PERMISSIONS
 from pyramid.view import view_config
 from pyramid_basemodel import Session
 from pyramid_bimt.models import AuditLogEntry
@@ -28,7 +29,10 @@ def audit_log(request):
     app_assets.need()
     table_assets.need()
 
-    new = AuditLogEntry.get_all(filter_by={'read': False}).count()
+    new = AuditLogEntry.get_all(
+        request=request,
+        filter_by={'read': False}
+    ).count()
     if new:  # pragma: no branch
         request.session.flash('{} new notifications.'.format(new))
 
@@ -70,7 +74,7 @@ class AuditLogAJAX(DatatablesDataView):
             entry.user.email,
         )
 
-        if self.request.user.admin:
+        if self.request.has_permission(ALL_PERMISSIONS):
             self.columns['action'] = """
             <a class="btn btn-xs btn-danger" href="{}">
               <span class="glyphicon glyphicon-remove-sign"></span> Delete